Move bitcoinLocktimeToNumber to bitcoin module
The function is closely related to bitcoin module.
